requirements.txt
# DockCheck — Dependencies
pyyaml>=6.0
docker>=7.0
jinja2>=3.1.2
pytest>=8.0
pytest-cov>=5.0
hypothesis>=6.0
dockerfile-parse>=2.0.0
ruamel.yaml>=0.17.0
